feat(passengers): implement booking passengers page with dynamic flight details
